+++ packages/adapter-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @d-dash/adapter-table
 
+## 0.1.3
+
+### Patch Changes
+
+- adjust style
+
 ## 0.1.2
 
 ### Patch Changes